Extract `JmesPathParenExpression` in its own file.
