+++ src/jmespath.net/Expressions/JmesParenExpression.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json.Linq;
+
+namespace DevLab.JmesPath.Expressions
+{
+    public class JmesParenExpression : JmesPathExpression
+    {
+        private readonly JmesPathExpression expression_;
+
+        public JmesParenExpression(JmesPathExpression expression)
+        {
+            expression_ = expression;
+        }
+
+        protected override JmesPathArgument Transform(JToken json)
+            => expression_.Transform(json);
+    }
+}
