+++ src/jmespath.net/Expressions/JmesPathOrExpression.cs
@@ -2,21 +2,6 @@
 
 namespace DevLab.JmesPath.Expressions
 {
-    public class JmesParenExpression : JmesPathExpression
-    {
-        private readonly JmesPathExpression expression_;
-
-        public JmesParenExpression(JmesPathExpression expression)
-        {
-            expression_ = expression;
-        }
-
-        protected override JmesPathArgument Transform(JToken json)
-        {
-            return expression_.Transform(json);
-        }
-    }
-
     public class JmesPathOrExpression : JmesPathCompoundExpression
     {
         /// <summary>
